requirements.txt
# Core AI
google-generativeai>=0.8.0
pyyaml>=6.0.1

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dateutil>=2.8.2
